Refactored setup, keep package self-contained
