+++ packages/mg-stripe/cli.js
@@ -1,0 +1,3 @@
+#! /usr/bin/env node --enable-source-maps
+
+import './build/bin/cli.js';
