+++ packages/mg-stripe/src/index.ts
@@ -1,5 +1,3 @@
-export {Logger} from './lib/Logger.js';
-export {Options} from './lib/Options.js';
-export {confirm} from './lib/commands/confirm.js';
-export {revert} from './lib/commands/revert.js';
-export {copy} from './lib/commands/copy.js';
+import command from './lib/command.js';
+
+export default command;
